Some errors should not stop parsing.
This probably is not a complete list but there is no way how to trigger
those I leave it as is with the infrastructure in place so we can add
them anytime later.

Couldn't come with a test cases as all the errors are impossible to
trigger if there is not a bug in snowcrash.