fix: skip invalid One Of entries
Fixes #699 by skipping invalid One Of entries when translating MSON to
API Elements.

In the example in #699, the parser (snowcrash) should skip the invalid
One Of even before it arrives in semantic analysis. Because we are planning
a complete rewrite of our parser, this issue was fixed during
semantic analysis (drafter).