fix(schema): produce minItems for empty array
`fixed-type` array without type would previous produce `anyOf` with an
empty array which is disallowed in [JSON Schema (2019-09
9.2.1.2)](https://json-schema.org/draft/2019-09/json-schema-core.html#rfc.section.9.2.1.2).
The array must be non-empty. Thus instead the change is to use
`maxItems` of `0` which indicates that the array must be empty. There is
no permitted types which may go inside the array.